refactored arm interface and added documentation
