--- src/arm_coordinator_interface/src/arm.c
@@ -1,1 +1,0 @@
-#include "arm.h"
