--- src/arm_coordinator_interface/srv/InputArmSetup.srv
@@ -1,4 +1,0 @@
-int32 ammount_color1
-int32 ammount_color2
----
-int32 ack
