--- src/arm_coordinator_interface/srv/NotifyArmFinished.srv
@@ -1,4 +1,0 @@
-int32 arm_type
-int32 ammount_moved_objects
----
-int32 ack
