--- src/arm_coordinator_interface/srv/NotifyObjectMovement.srv
@@ -1,7 +1,0 @@
-int32 turtle_id
-int32 obj_diff
-string object_color
----
-int32 ack
-
-# ack = 1 if turtle left, ack = 0 if turtle stayed, else ack < 0 error
