+++ src/arm_interface/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.8)
-project(arm_coordinator_interface)
+project(arm_interface)
 
 if(CMAKE_COMPILER_IS_GNUCXX OR CMAKE_CXX_COMPILER_ID MATCHES "Clang")
   add_compile_options(-Wall -Wextra -Wpedantic)
@@ -10,12 +10,8 @@ find_package(ament_cmake REQUIRED)
 find_package(builtin_interfaces REQUIRED)
 find_package(rosidl_default_generators REQUIRED)
 
-rosidl_generate_interfaces(arm_coordinator_interface
-  "srv/NotifyObjectMovement.srv"
-  "srv/NotifyArmFinished.srv"
-  "srv/InputArmSetup.srv"
-)
-
+#rosidl_generate_interfaces(arm_coordinator_interface
+#)
 
 if(BUILD_TESTING)
   find_package(ament_lint_auto REQUIRED)
@@ -23,34 +19,35 @@ if(BUILD_TESTING)
   set(ament_cmake_cpplint_FOUND TRUE)
   ament_lint_auto_find_test_dependencies()
 endif()
+
 #
 # Create the library target
-add_library(arm_coordinator_interface_lib
-  src/arm.c
+add_library(arm_interface_lib
+  src/arm_interface.cpp
 )
 
 # Use target_include_directories to specify public include paths.
-target_include_directories(arm_coordinator_interface_lib PUBLIC
-  $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include/arm_coordinator_interface>
-  $<INSTALL_INTERFACE:include/arm_coordinator_interface>
+target_include_directories(arm_interface_lib PUBLIC
+  $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include/arm_interface>
+  $<INSTALL_INTERFACE:include/arm_interface>
 )
 
 # Install the library target and export it.
 install(TARGETS
-  arm_coordinator_interface_lib
-  EXPORT arm_coordinator_interfaceTargets
+  arm_interface_lib
+  EXPORT arm_interfaceTargets
   LIBRARY DESTINATION lib
   ARCHIVE DESTINATION lib
   RUNTIME DESTINATION bin
 )
 
 # Install header files.
-install(DIRECTORY include/arm_coordinator_interface/
-  DESTINATION include/arm_coordinator_interface
+install(DIRECTORY include/arm_interface/
+  DESTINATION include/arm_interface
 )
 
 # Export the targets so that they can be found by other packages.
-ament_export_targets(arm_coordinator_interfaceTargets HAS_LIBRARY_TARGET)
+ament_export_targets(arm_interfaceTargets HAS_LIBRARY_TARGET)
 ament_export_dependencies(builtin_interfaces)
 
 ament_package()
