+++ src/arm_interface/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>arm_coordinator_interface</name>
+  <name>arm_interface</name>
   <version>0.0.0</version>
   <description>TODO: Package description</description>
   <maintainer email="andresilvatelles@gmail.com">andre</maintainer>
@@ -13,7 +13,6 @@
   <member_of_group>rosidl_interface_packages</member_of_group>
 
   <depend>builtin_interfaces</depend>
-
 
   <test_depend>ament_lint_auto</test_depend>
   <test_depend>ament_lint_common</test_depend>
