+++ src/coordinator/package.xml
@@ -8,8 +8,9 @@
   <license>Apache-2.0</license>
 
   <buildtool_depend>ament_cmake</buildtool_depend>
-  <depend>arm_coordinator_interface</depend>
-  <depend>turtle_coordinator_interface</depend>
+
+  <depend>arm_interface</depend>
+  <depend>turtle_interface</depend>
   <depend>resource_manager_interface</depend>
 
   <depend>rosidl_default_generators</depend>
